pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reklaim-whatsapp-ai"
version = "0.1.0"
description = "WhatsApp AI microservice using FastAPI, Celery, and LangGraph with AWS Bedrock"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "langgraph>=0.3",
    "langgraph-checkpoint-redis>=0.1",
    "langchain-aws>=0.2",
    "langchain-core>=0.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "moto[bedrock]>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
